Moving killswitch RC test to RC suite
This is done because during new remote config testing, a full build repo pipeline is done, testing the local config files. Any test that tries to download remote config from cdn as a test condition will not work with these local tests. Instead of trying to exclude one test for the killswitch test suite under specific circumstances, it is easier to move the test to the remote_config tests file which does not work and is not run in this situation.
