+++ test/qa/test_remote_config.py
@@ -3,6 +3,7 @@ import json
 import subprocess
 import time
 import uuid
+import os
 
 import pytest
 import sh
@@ -18,6 +19,7 @@ from lib.daemon import (
     set_rc_retry_custom_time,
 )
 from constants import RC_TIMEOUT
+from test_killswitch import MSG_KILLSWITCH_ON
 
 RC_INVALID_VERSION_MESSAGE = "invalid version constraint: improper constraint: y"
 RC_REMOTE_MESSAGES = [
@@ -774,3 +776,59 @@ def test_remote_config_change_local_meshnet_config_settings(
             sh.nordvpn("meshnet")
         except sh.ErrorReturnCode:
             pytest.fail(reason="nordvpn meshnet is not enabled")
+
+
+def test_killswitch_enabled_does_not_affect_cdn_with_firewall_mark(
+        initialized_app_with_remote_config # noqa: ARG001
+        ):
+    """
+    Test for a scenario where despite killswitch being enabled the network is still accessible for remote config.
+
+    Test steps:
+        1. Set up required environment variables
+        2. Enable killswitch
+        3. Remove previously fetched config files
+        4. Attempt to fetch remote config
+        5. Verify that the config is fetched successfully
+
+    Jira ID: LVPN-8626
+    """
+
+    conf_dir = "/var/lib/nordvpn/conf/"
+    expected_files = [
+        "libtelio.json",
+        "meshnet.json",
+    ]
+    def printdir():
+        os.system(f"cd {conf_dir}/../ && ls -la && ls {conf_dir} -a")
+
+    # enabling killswitch should not affect http transport of the remote config
+    assert MSG_KILLSWITCH_ON in sh.nordvpn.set.killswitch("on")
+    assert daemon.is_killswitch_on()
+
+    sh.nordvpn.disconnect()
+    assert daemon.is_disconnected()
+    print("--- before manual files removal")
+    printdir()
+    # remove previously fetched files
+    # upon restart, they should be loaded again
+    os.system(f"sudo rm -rf {conf_dir}")
+
+    # make sure the rc files are gone
+    for fname in expected_files:
+        path = os.path.join(conf_dir, fname)
+        res = os.popen(f"sudo test -f {path} && echo exists || echo missing").read().strip()
+        assert res == "missing", f"File {path} should not exist"
+
+    daemon.restart()
+    print("--- after restart")
+    printdir()
+
+    for _ in range(10):
+        time.sleep(1)
+        printdir()
+
+    for fname in expected_files:
+        path = os.path.join(conf_dir, fname)
+        res = os.popen(f"sudo test -f {path} && echo exists || echo missing").read().strip()
+        assert res == "exists", f"File {os.path} should exist after kill-switch was enabled"
